REFACTOR: move to SampleLayer and change to iid from string used in router complete
