+++ db/python/layers/sample.py
@@ -1,9 +1,10 @@
-from typing import Dict, List, Any, Optional
+from typing import Dict, List, Any, Optional, Union
 
 from pydantic import BaseModel
 
 from db.python.connect import NotFoundError
 from db.python.layers.base import BaseLayer, Connection
+from db.python.layers.sequence import SampleSequenceLayer, SequenceUpsert
 from db.python.tables.project import ProjectId
 from db.python.tables.sample import SampleTable
 from db.python.tables.sequence import SampleSequencingTable
@@ -12,7 +13,6 @@ from models.enums import SampleType
 from models.models.sample import (
     Sample,
     sample_id_format_list,
-    sample_id_transform_to_raw,
 )
 
 
@@ -24,6 +24,19 @@ class SampleUpsert(BaseModel):
     meta: Optional[Dict] = {}
     participant_id: Optional[int] = None
     active: Optional[bool] = None
+
+
+class SampleBatchUpsertItem(SampleUpsert):
+    """Update model for sample with sequences list"""
+
+    id: Optional[Union[str, int]]
+    sequences: List[SequenceUpsert]
+
+
+class SampleBatchUpsert(BaseModel):
+    """Upsert model for batch Samples"""
+
+    samples: List[SampleBatchUpsertItem]
 
 
 class SampleLayer(BaseLayer):
@@ -205,7 +218,7 @@ class SampleLayer(BaseLayer):
 
         # Otherwise update
         internal_id = await self.update_sample(
-            id_=sample_id_transform_to_raw(sample.id),
+            id_=sample.id,
             meta=sample.meta,
             participant_id=sample.participant_id,
             type_=sample.type,
@@ -248,3 +261,17 @@ class SampleLayer(BaseLayer):
             )
 
         return rows
+
+    async def batch_upsert_samples(
+        self, samples: SampleBatchUpsert, seqt: SampleSequenceLayer
+    ):
+        """Batch upsert a list of samples with sequences"""
+        # Create or update samples
+        iids = [await self.upsert_sample(s) for s in samples.samples]
+
+        # Upsert all sequences with paired sids
+        sequences = zip(iids, [x.sequences for x in samples.samples])
+        seqs = [await seqt.upsert_sequences(iid, seqs) for iid, seqs in sequences]
+
+        # Format and return response
+        return dict(zip(iids, seqs))
